# docs/openapi.yml
openapi: 3.1.0
info:
  title: Tracecat API
  summary: Tracecat API
  description: Tracecat is the open source Tines / Splunk SOAR alternative. You can
    operate Tracecat in headless mode by using the API to create, manage, and run
    workflows.
  termsOfService: https://docs.google.com/document/d/e/2PACX-1vQvDe3SoVAPoQc51MgfGCP71IqFYX_rMVEde8zC4qmBCec5f8PLKQRdxa6tsUABT8gWAR9J-EVs2CrQ/pub
  contact:
    name: Tracecat Founders
    email: founders@tracecat.com
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  version: '1'
paths:
  /webhooks/{workflow_id}/{secret}:
    post:
      tags:
      - public
      summary: Incoming Webhook
      description: 'Webhook endpoint to trigger a workflow.


        This is an external facing endpoint is used to trigger a workflow by sending
        a webhook request.

        The workflow is identified by the `path` parameter, which is equivalent to
        the workflow id.'
      operationId: public-incoming_webhook
      parameters:
      - name: secret
        in: path
        required: true
        schema:
          type: string
          title: Secret
      - name: workflow_id
        in: path
        required: true
        schema:
          anyOf:
          - type: string
            pattern: wf_[0-9a-zA-Z]+
          - type: string
            pattern: wf-[0-9a-f]{32}
          title: Workflow Id
      - name: echo
        in: query
        required: false
        schema:
          type: boolean
          description: Echo the request payload back to the caller
          default: false
          title: Echo
        description: Echo the request payload back to the caller
      - name: content-type
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Content-Type
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionCreateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /webhooks/{workflow_id}/{secret}/wait:
    post:
      tags:
      - public
      summary: Incoming Webhook Wait
      description: 'Webhook endpoint to trigger a workflow.


        This is an external facing endpoint is used to trigger a workflow by sending
        a webhook request.

        The workflow is identified by the `path` parameter, which is equivalent to
        the workflow id.'
      operationId: public-incoming_webhook_wait
      parameters:
      - name: secret
        in: path
        required: true
        schema:
          type: string
          title: Secret
      - name: workflow_id
        in: path
        required: true
        schema:
          anyOf:
          - type: string
            pattern: wf_[0-9a-zA-Z]+
          - type: string
            pattern: wf-[0-9a-f]{32}
          title: Workflow Id
      - name: content-type
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Content-Type
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Public-Incoming Webhook Wait
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /webhooks/{workflow_id}/{secret}/interactions/{category}:
    post:
      tags:
      - public
      summary: Receive Interaction
      description: "Process incoming workflow interactions from external services.\n\
        \nArgs:\n    workflow_id: ID of the workflow to interact with\n    payload:\
        \ Raw interaction payload from the external service\n    category: Category\
        \ of interaction (e.g., slack)\n\nReturns:\n    Response confirming the interaction\
        \ was processed\n\nRaises:\n    HTTPException: If payload is invalid or workflow\
        \ interaction fails"
      operationId: public-receive_interaction
      parameters:
      - name: category
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/InteractionCategory'
      - name: secret
        in: path
        required: true
        schema:
          type: string
          title: Secret
      - name: workflow_id
        in: path
        required: true
        schema:
          anyOf:
          - type: string
            pattern: wf_[0-9a-zA-Z]+
          - type: string
            pattern: wf-[0-9a-f]{32}
          title: Workflow Id
      - name: content-type
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Content-Type
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiveInteractionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /workflows/{workflow_id}/webhook:
    post:
      tags:
      - triggers
      summary: Create Webhook
      description: Create a webhook for a workflow.
      operationId: triggers-create_webhook
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workflow_id
        in: path
        required: true
        schema:
          anyOf:
          - type: string
            pattern: wf_[0-9a-zA-Z]+
          - type: string
            pattern: wf-[0-9a-f]{32}
          title: Workflow Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - triggers
      summary: Get Webhook
      description: Get the webhook from a workflow.
      operationId: triggers-get_webhook
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workflow_id
        in: path
        required: true
        schema:
          anyOf:
          - type: string
            pattern: wf_[0-9a-zA-Z]+
          - type: string
            pattern: wf-[0-9a-f]{32}
          title: Workflow Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - triggers
      summary: Update Webhook
      description: Update the webhook for a workflow. We currently supprt only one
        webhook per workflow.
      operationId: triggers-update_webhook
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workflow_id
        in: path
        required: true
        schema:
          anyOf:
          - type: string
            pattern: wf_[0-9a-zA-Z]+
          - type: string
            pattern: wf-[0-9a-f]{32}
          title: Workflow Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookUpdate'
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    InteractionCategory:
      type: string
      enum:
      - slack
      title: InteractionCategory
...
    ReceiveInteractionResponse:
      properties:
        message:
          type: string
          title: Message
      type: object
      required:
      - message
      title: ReceiveInteractionResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WebhookCreate:
      properties:
        status:
          $ref: '#/components/schemas/WebhookStatus'
          default: offline
        method:
          $ref: '#/components/schemas/WebhookMethod'
          default: POST
        entrypoint_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Entrypoint Ref
      type: object
      title: WebhookCreate
    WebhookMethod:
      type: string
      enum:
      - GET
      - POST
    WebhookRead:
      properties:
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        owner_id:
          type: string
          format: uuid
          title: Owner Id
        id:
          type: string
          title: Id
        secret:
          type: string
          title: Secret
        status:
          $ref: '#/components/schemas/WebhookStatus'
        entrypoint_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Entrypoint Ref
        filters:
          type: object
          title: Filters
        method:
          $ref: '#/components/schemas/WebhookMethod'
        workflow_id:
          type: string
          title: Workflow Id
        url:
          type: string
          title: Url
      type: object
      required:
      - owner_id
      - id
      - secret
      - status
      - filters
      - method
      - workflow_id
      - url
      title: WebhookRead
    WebhookStatus:
      type: string
      enum:
      - online
      - offline
    WebhookUpdate:
      properties:
        status:
          anyOf:
          - $ref: '#/components/schemas/WebhookStatus'
          - type: 'null'
        method:
          anyOf:
          - $ref: '#/components/schemas/WebhookMethod'
          - type: 'null'
        entrypoint_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Entrypoint Ref
      type: object
      title: WebhookUpdate
...
    WorkflowExecutionCreateResponse:
      properties:
        message:
          type: string
          title: Message
        wf_id:
          type: string
          title: Wf Id
        wf_exec_id:
          type: string
          pattern: (?P<workflow_id>wf-[0-9a-f]{32}|wf_[0-9a-zA-Z]+)[:/](?P<execution_id>(exec_[0-9a-zA-Z]+|exec-[\w-]+|sch-[0-9a-f]{32}-.*))
          title: Wf Exec Id
        payload:
          title: Payload
      type: object
      required:
      - message
      - wf_id
      - wf_exec_id
      title: WorkflowExecutionCreateResponse
...
  securitySchemes:
    APIKeyCookie:
      type: apiKey
      in: cookie
      name: fastapiusersauth
    APIKeyHeader:
      type: apiKey
      in: header
      name: x-tracecat-service-key
...